code_pancreas/requirements.txt
tensorboardx
pyyaml
yacs
einops
SimpleITK
pandas
numpy
scikit-image
sklearn
matplotlib
torchio
medpy
h5py
geoopt
